+++ fabric/src/main/java/com/avp/fabric/data/compatibility/common/CommonConstants.java
@@ -47,6 +47,8 @@ public class CommonConstants {
 
     public static final TagKey<Item> ORES = createItem("ores");
 
+    public static final TagKey<Item> RAW_MATERIALS_ALUMINUM = createItem("raw_materials/aluminum");
+
     public static final TagKey<Item> URANIUM = createItem("ingots/uranium");
 
     /*